+++ domain/mocks/CharacterWriteRepository.go
@@ -14,11 +14,11 @@ type CharacterWriteRepository struct {
 }
 
 // StoreByID provides a mock function with given fields: ctx, id
-func (_m *CharacterWriteRepository) StoreByID(ctx context.Context, id int64) error {
+func (_m *CharacterWriteRepository) StoreByID(ctx context.Context, id int) error {
 	ret := _m.Called(ctx, id)
 
 	var r0 error
-	if rf, ok := ret.Get(0).(func(context.Context, int64) error); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, int) error); ok {
 		r0 = rf(ctx, id)
 	} else {
 		r0 = ret.Error(0)
